# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CandidateRefreshResponse": {
        "properties": {
          "candidate_count": {
            "title": "Candidate Count",
            "type": "integer"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "provider_failure_count": {
            "title": "Provider Failure Count",
            "type": "integer"
          },
          "provider_request_count": {
            "title": "Provider Request Count",
            "type": "integer"
          },
          "status": {
            "enum": [
              "fresh",
              "partial",
              "stale",
              "no_candidates"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "candidate_count",
          "provider_request_count",
          "provider_failure_count",
          "duration_ms"
        ],
        "title": "CandidateRefreshResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecommendationFeedbackResponse": {
        "properties": {
          "canonical_entity_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Entity Id"
          },
          "feedback_type": {
            "enum": [
              "LIKE",
              "DISLIKE",
              "NOT_INTERESTED"
            ],
            "title": "Feedback Type",
            "type": "string"
          },
          "provider_identity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderEntityIdentity"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "const": "saved",
            "default": "saved",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "feedback_type"
        ],
        "title": "RecommendationFeedbackResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "R2 API for a canonical library, playback, and personal music intelligence.",
    "title": "MusicScope V2 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/recommendations/candidates/refresh": {
      "post": {
        "operationId": "refresh_candidates_api_v1_recommendations_candidates_refresh_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateRefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Candidates",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/discover": {
      "get": {
        "operationId": "discover_recommendations_api_v1_recommendations_discover_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "default": "for-you",
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 60,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverRecommendationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover Recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/external/{provider}/tracks/{provider_id}": {
      "get": {
        "operationId": "external_track_detail_api_v1_recommendations_external__provider__tracks__provider_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "External Track Detail",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/external/{provider}/tracks/{provider_id}/playback-source": {
      "post": {
        "operationId": "external_track_playback_api_v1_recommendations_external__provider__tracks__provider_id__playback_source_post",
        "parameters": [
          {
            "in": "path",
            "name": "provider",
            "required": true,
            "schema": {
              "title": "Provider",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaybackSourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "External Track Playback",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/feedback": {
      "post": {
        "operationId": "save_recommendation_feedback_api_v1_recommendations_feedback_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationFeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationFeedbackResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Recommendation Feedback",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/home": {
      "get": {
        "operationId": "home_recommendations_api_v1_recommendations_home_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeRecommendationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Home Recommendations",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/profile": {
      "get": {
        "operationId": "recommendation_profile_api_v1_recommendations_profile_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recommendation Profile",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/profile/rebuild": {
      "post": {
        "operationId": "rebuild_recommendation_profile_api_v1_recommendations_profile_rebuild_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileRebuildResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rebuild Recommendation Profile",
        "tags": [
          "recommendations"
        ]
      }
    },
    "/api/v1/recommendations/settings": {
      "patch": {
        "operationId": "update_recommendation_settings_api_v1_recommendations_settings_patch",
        "parameters": [
          {
            "in": "header",
            "name": "x-musicscope-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Musicscope-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecommendationSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Recommendation Settings",
        "tags": [
          "recommendations"
        ]
      }
    },
...
  }
}